Fixed circular dependency failing test (152)
