+++ src/services/PollingService.ts
@@ -1,4 +1,5 @@
 import type { Disposable } from '../types';
+import { withResolvers, type PromiseWithCancel } from '../util';
 
 export type Runner = () => Promise<void>;
 
@@ -54,5 +55,65 @@ export class PollingService implements Disposable {
 
   dispose = async (): Promise<void> => {
     this.stop();
+  };
+}
+
+/**
+ * Call the given poll function at an interval.
+ * - If the poll result resolves to true, stop polling and resolve the promise.
+ * - If the poll function throws, stop polling and reject the promise.
+ * @param poll
+ * @param intervalMs
+ * @param timeoutMs
+ * @returns Promise that resolves when the poll function returns true + a `reject`
+ * function that can be used to cancel the polling.
+ */
+export function pollUntilTrue(
+  poll: () => Promise<boolean>,
+  intervalMs: number,
+  timeoutMs?: number
+): PromiseWithCancel<true> {
+  const { promise, resolve, reject } = withResolvers<true>();
+
+  let timeout: NodeJS.Timeout;
+  const poller = new PollingService();
+
+  /** Stop polling and resolve / reject promise */
+  function resolveOrReject(trueOrError: true | Error): void {
+    poller.stop();
+    clearTimeout(timeout);
+
+    if (trueOrError === true) {
+      resolve(trueOrError);
+    } else {
+      reject(trueOrError);
+    }
+  }
+
+  /** Cancel polling */
+  const cancel = (): void => {
+    resolveOrReject(new Error('Polling cancelled'));
+  };
+
+  if (timeoutMs != null) {
+    timeout = setTimeout(() => {
+      cancel();
+    }, timeoutMs);
+  }
+
+  poller.start(async () => {
+    try {
+      const isTrue = await poll();
+      if (isTrue) {
+        resolveOrReject(true);
+      }
+    } catch (err) {
+      resolveOrReject(err instanceof Error ? err : new Error(String(err)));
+    }
+  }, intervalMs);
+
+  return {
+    promise,
+    cancel,
   };
 }
